% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_contenido_propio.R
\name{calcula_contenido_propio}
\alias{calcula_contenido_propio}
\title{calcula_contenido_propio}
\usage{
calcula_contenido_propio(df)
}
\arguments{
\item{df}{Dataframe con columnas: is_quote}
}
\description{
Esta es una funcion pra calcular si el contenido es propio o retweet o quote
}
\details{
Esta funcion calcula si el contenido es propio o si es retweet
}
\examples{
calcula_contenido_propio(df_tweets_from_rtweet)
}
