% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_frecuencia_palabras.R
\name{grafica_frecuencia_palabras}
\alias{grafica_frecuencia_palabras}
\title{grafica_frecuencia_palabras}
\usage{
grafica_frecuencia_palabras(
  corp_tmln,
  name = "nombre",
  type = "usuario_tendencia",
  num_freq = 10,
  folder = ".",
  stopwords_pers = ""
)
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{name}{nombre del usuario o del trend a analizar}

\item{type}{el tipo de objeto twitter a analizas. Puede ser "tendencia" o "usuario"}

\item{num_freq}{Top n de palabras a mostrar}

\item{folder}{Folder donde se va  a guardar la imagen}

\item{stopword_pers}{lista de stopwords personalizadas para filtrar}
}
\description{
Función que grafica la frecuencia de palabras usadas
}
\details{
Genera wordcloud quitando hashtags y usuarios
}
\examples{
grafica_frecuencia_palabras(corp_tmln,"covid19","tendencia",20,"./03Graficos/")
}
