% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_red_hastags.R
\name{grafica_hashtags}
\alias{grafica_hashtags}
\title{grafica_hashtags}
\usage{
grafica_hashtags(
  corp_tmln,
  name = "nombre",
  type = "usuario_tendencia",
  num_freq = 25,
  folder = ".",
  stopwords_pers = ""
)
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{name}{nombre del usuario o del trend a analizar}

\item{type}{el tipo de objeto twitter a analizas. Puede ser "tendencia" o "usuario"}

\item{num_freq}{Top n de palabras a mostrar}

\item{folder}{Folder donde se va  a guardar la imagen}

\item{stopword_pers}{lista de stopwords personalizadas para filtrar}
}
\description{
Función que grafica una red de hastags
}
\details{
Genera grafica de barras
}
\examples{
grafica_hashtags('nerudista',corpus,"./03Graficos/")
}
