% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_tipo_contenido.R
\name{grafica_tipo_contenido}
\alias{grafica_tipo_contenido}
\title{grafica_contenido}
\usage{
grafica_tipo_contenido(tmln, folder)
}
\arguments{
\item{tmln}{dataframe con una columna llamada "contenido_propio"}

\item{folder}{Folder donde se va  a guardar la imagen}
}
\description{
Función que grafica un plot de barras
}
\details{
Genera grafica de barras
}
\examples{
grafica_contenido('tmln')
}
