% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_wordcloud_clean.R
\name{grafica_wordcloud_clean}
\alias{grafica_wordcloud_clean}
\title{grafica_wordcloud_clean}
\usage{
grafica_wordcloud_clean(user, corp_tmln, folder)
}
\arguments{
\item{user}{corpus que será tokenizado}

\item{corp_tmln}{corpus que será tokenizado}

\item{folder}{Folder donde se va  a guardar la imagen}
}
\description{
Función que grafica una nube de palabras sin hastags ni usuarios
}
\details{
Genera wordcloud quitando hashtags y usuarios
}
\examples{
grafica_wordcloud_clean('nerudista',corpus,"./03Graficos/")
}
