% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_usuario_rtweet.R
\name{leer_usuario_rtweet}
\alias{leer_usuario_rtweet}
\title{leer_usuario_rtweet}
\usage{
leer_usuario_rtweet(user, num)
}
\arguments{
\item{user}{Usuario del que se leerá el timeline de Twitter}

\item{num}{Número de tuits a leer del timeline}
}
\description{
Esta es una funcion de ejemplo para multiplicar un valor o vector \code{x} por dos.
}
\details{
Esta funcion lee el timeline y calcula si el contenido es propio o si es retweet
}
\examples{
leer_usuario_rtweet('nerudista',3)
}
