% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-stopwords.R
\name{stopwords}
\alias{stopwords}
\title{Stopwords}
\usage{
stopwords(language = "en", source = "snowball")
}
\description{
Return a character vector of stopwords.
See \code{stopwords::\link[stopwords:stopwords]{stopwords()}} for details.
}
