% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizar_corpus.R
\name{tokenizar_corpus}
\alias{tokenizar_corpus}
\title{tokenizar_tweets}
\usage{
tokenizar_corpus(corp_tmln, stopwords_pers = "")
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{stopword_pers}{lista de stopwords personalizadas para filtrar}
}
\description{
Función que tokenizaun corpus y quita stopwords
}
\details{
Genera un objeto tokens
}
\examples{
grafica_frecuencia_palabras(corp_tmln,"covid19","tendencia",20,"./03Graficos/")
}
