% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_contenido_propio.R
\name{calcula_contenido_propio}
\alias{calcula_contenido_propio}
\title{calcula_contenido_propio}
\usage{
calcula_contenido_propio(df = df_tweets_from_rtweet)
}
\arguments{
\item{df}{Dataframe con columnas: is_quote, is_retweet y reply_to_screen_name.}
}
\description{
Esta es una funcion para calcular si el contenido es propio o retweet o quote
}
\details{
Esta funcion calcula si el contenido es propio o si es retweet. Las tres son obligatorias en el dataframe
}
\examples{
## Obtener un dataframe con rtweet
df_tweets <- rtweet::get_timeline(user="usuarioX",n = 600)

## Generar columna contenido_propio
calcula_contenido_propio(df=df_tweets)

}
