% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_frecuencia_palabras.R
\name{grafica_frecuencia_palabras}
\alias{grafica_frecuencia_palabras}
\title{grafica_frecuencia_palabras}
\usage{
## grafica_frecuencia_palabras(
corp_tmln=corpus,
name="nombre",
type="usuario_tendencia",
num_freq = 10,
folder=".",
stopwords_pers=c("una","palabra","otra")
)
}
\arguments{
\item{corp_tmln}{Corpus que ser치 tokenizado}

\item{name}{Nombre del usuario o del trend a analizar}

\item{type}{El tipo de objeto twitter a analizar. Puede ser "tendencia" o "usuario". Otros valores causan error.}

\item{num_freq}{Top n de palabras a mostrar}

\item{folder}{Folder donde se va  a guardar la imagen}

\item{stopword_pers}{Vector de  stopwords personalizadas para filtrar}
}
\description{
Funci칩n que genera una gr치fica de barras con la frecuencia de palabras usadas en el corpus.
}
\details{
Genera wordcloud quitando hashtags y usuarios
}
\examples{
## Generar corpus
corpus_tweets <- df_tweets \%>\% select(text) \%>\% corpus()

## Generar gr치fica
grafica_frecuencia_palabras(corpus_tweets,"covid19","tendencia",20,"./03Graficos/")
}
