% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_red_hastags.R
\name{grafica_hashtags}
\alias{grafica_hashtags}
\title{grafica_hashtags}
\usage{
grafica_hashtags(user, corp_tmln, folder)
}
\arguments{
\item{user}{corpus que será tokenizado}

\item{corp_tmln}{corpus que será tokenizado}

\item{folder}{Folder donde se va  a guardar la imagen}
}
\description{
Función que grafica una red de hastags
}
\details{
Genera grafica de barras
}
\examples{
grafica_hashtags('nerudista',corpus,"./03Graficos/")
}
