% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_red_hastags.R
\name{grafica_hashtags}
\alias{grafica_hashtags}
\title{grafica_hashtags}
\usage{
## grafica_hashtags(
corp_tmln=corpus,
num_freq = 10,
stopwords_pers=c("una","palabra","otra")
)
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{num_freq}{Número de palabras a mostrar.}

\item{stopword_pers}{Vector de stopwords personalizadas para filtrar}
}
\description{
Función que grafica una red de hastags a partir de un corpus recibido.
}
\details{
Genera grafica red con los hashtags más representativos del corpus recibido.
}
\examples{
## Generar corpus
corpus_tweets <- df_tweets \%>\% select(text) \%>\% corpus()

## Generar gráfica
grafica_hashtags(corpus_tweets,"covid19","tendencia",20)

## Para guardar
png(file=paste0( folder,"/all_wordcloud_top.png"),width=600,height=600, units='px', res=120)
grafica_hashtags(corpus_tweets,"covid19","tendencia",20)
dev.off()


}
