% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_red_usuarios.R
\name{grafica_usuarios}
\alias{grafica_usuarios}
\title{grafica_usuarios}
\usage{
grafica_usuarios(user, corp_tmln, folder)
}
\arguments{
\item{user}{corpus que será tokenizado}

\item{corp_tmln}{corpus que será tokenizado}

\item{folder}{Folder donde se va  a guardar la imagen}
}
\description{
Función que grafica una red de usuarios
}
\details{
Genera grafica de barras
}
\examples{
grafica_usuarios('nerudista',corpus,"./03Graficos/")
}
