% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_red_usuarios.R
\name{grafica_usuarios}
\alias{grafica_usuarios}
\title{grafica_usuarios}
\usage{
## grafica_usuarios(
corp_tmln=corpus,
name="nombre",
type="usuario_tendencia",
num_freq = 10,
folder=".",
stopwords_pers=c("una","palabra","otra")
)
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{name}{Nombre del usuario o del trend a analizar}

\item{type}{El tipo de objeto twitter a analizar. Puede ser "tendencia" o "usuario". Otros valores pueden causar error}

\item{num_freq}{Número de palabras a mostrar.}

\item{folder}{Folder donde se va  a guardar la imagen}

\item{stopword_pers}{Vector de stopwords personalizadas para filtrar}
}
\description{
Función que grafica una red de usuarios de Twitter a partir de un corpus recibido.
}
\details{
Genera grafica de red con los usuarios más representativos del corpus recibido.
}
\examples{
## Generar corpus
corpus_tweets <- df_tweets \%>\% select(text) \%>\% corpus()

## Generar gráfica
grafica_usuarios(corpus_tweets,"covid19","tendencia",20,"./03Graficos/")
}
