% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_wordcloud_all.R
\name{grafica_wordcloud_all}
\alias{grafica_wordcloud_all}
\title{grafica_wordcloud_all}
\usage{
## grafica_wordcloud_clean(
corp_tmln=corpus,
name="nombre",
type="usuario_tendencia",
num_freq = 10,
stopwords_pers=c("una","palabra","otra")
)
}
\arguments{
\item{corp_tmln}{corpus que será tokenizado}

\item{name}{Nombre del usuario o del trend a analizar}

\item{type}{El tipo de objeto twitter a analizar. Puede ser "tendencia" o "usuario". Otros valores pueden causar error}

\item{num_freq}{Número de palabras a mostrar.}

\item{stopword_pers}{Vector de stopwords personalizadas para filtrar}
}
\description{
Función que grafica una nube de palabras incluyendo hastags y nombres de usuarios
}
\details{
Genera grafica de nube de palabras con hastags y nombres de usuarios
}
\examples{
## Generar corpus
corpus_tweets <- df_tweets \%>\% select(text) \%>\% corpus()

## Generar gráfica
grafica_wordcloud_all(corpus_tweets,"covid19","tendencia",20)
}
