% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizar_corpus.R
\name{tokenizar_corpus}
\alias{tokenizar_corpus}
\title{tokenizar_tweets}
\usage{
tokenizar_corpus(corp_tmln= corpus_twitter
stopwords_pers= c("una","palabra","otra"))
}
\arguments{
\item{corp_tmln}{Corpus que será tokenizado}

\item{stopword_pers}{Vettor de stopwords personalizadas para filtrar}
}
\description{
Función que tokeniza un corpus y quita stopwords
}
\details{
Genera un objeto tokens
}
